requirements.txt
depthai>=2.21.0
opencv-python>=4.7.0
pillow>=9.0.0
pyserial>=3.5
pynmea2>=1.18.0
numpy>=1.23.0